#7 add a auth middleware
